+++ starsim/__init__.py
@@ -21,7 +21,7 @@ import sciris as sc
 root = sc.thispath(__file__).parent
 
 # Start imports: version and settings
-t('settings') # SLow since import Numba
+t('settings') # Slow since importing Numba
 from .version import __version__, __versiondate__, __license__
 from .settings import dtypes, options, style, load_fonts
 
@@ -40,7 +40,7 @@ sc.require(reqs=['sciris>=3.2.8', 'pandas>=2.0.0'], die=False,
 t('utils')
 from .utils import (
     ndict, warn, find_contacts, standardize_netkey, parse_age_range, apply_age_range,
-    standardize_data, validate_sim_data, load, save, plot_args, show,
+    standardize_data, validate_sim_data, load, save, shrink, plot_args, show,
     return_fig,
 )